Adding service dependency viewing methods
So, we can compute an array-encoded "tree" of service dependents/depends_on data from
a given service.  There is now a helper (display_tree) to turn this (currently) into
a simple ul/li structure for rendering (this won't be sufficient for long, but it's at
least a starting point).  We display the depends_on tree on the service show view now.

Granted, this is implementing a graph-y library indirectly, but I'm not ready to
refactor this any time soon, until I know how it's going to flesh out.