Breaking linkage between Service and DeployedService.
Since Services are free to change, and free to change in their relationships to
other services, but may still be deployed, a linkage between DeployedService and
Service is misleading at best.  If we want to find "instances of the Postgres service
which are deployed" we'll probably add the ability to search DeployedService by
service_name == 'Postgres'.