Fixing whack-ass spec breakage on deployment.
First off, I'm in India for a bit, and my laptop is on Indian time.  I've been
developing this crap for a while now and finally deployed it out to one of our
slices, which is on one of the US time zones (I forget which).

Anyway, there was a bug that showed up (unrelated to this fix, related to
rails route globbing).  I ran the specs on the slice and got 36 failures.  The
slice is running old-ass ruby 1.8.5 so I went down that rabbit hole.  I had
vinbarnes run it on his 1.8.6-but-not-the-latest version and he had the failures
also.  I ran the specs on a local ubuntu-under-vmware installation and they worked
fine.

A bunch of rabbit-holes later, I was messing around on script/console and saw that
the times I should've been seeing weren't lining up.  Whoops.  Time.now is local, and
the times coming back from the database were time-zone specific.

So then I dug around and found this:

  http://marklunds.com/articles/one/402

Oh, right.  So the named scope fix.  Then the other bit of stupidity was the default
start_time in the Deployment exemplar.  Doesn't matter, right?  Yeah, unless you want
all Deployments to be non-current, but the specs are expecting current Deployments
in a lot of cases, so fix that stupidity.

Yay.