One more time-zone related tweak.
Nice.  If you want a ::Date:: format (i.e., use the date class when declaring the formatter)
it puts the formatter into the same data structure as if you specify it via ::Time::, but
when you use it, it'll do wacky crap with the time part of the string.  If you specify it
via ::Time:: it will behave.  So, (a) why is this undocumented, and (b) why does the various
documentation seem to prefer using the bitched-up ::Date:: route?  Maybe someone was too busy
browsing threadless.com for a new ironic t-shirt instead of thinking about the API.

Also, made the test checking this a bit more robust.